Delete tf se_atten_v2
